- refactor components - restructure project files
