+++ src/components/Cockpit/Cockpit.css
@@ -1,0 +1,29 @@
+.red {
+    color: red;
+  }
+  
+  .bold {
+    font-weight: bold;
+  }
+  
+  .Cockpit button {
+    border: 1px solid blue;
+    padding: 16px;
+    background-color: green;
+    color: white;
+    cursor: pointer;
+  }
+  
+  .Cockpit button:hover {
+    background-color: lightgreen;
+    color: black;
+  }
+  
+  .Cockpit button.Red {
+    background-color: red;
+  }
+  
+  .Cockpit button.Red:hover {
+    background-color: salmon;
+    color: black;
+  }
