+++ src/components/Persons/Persons.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import Person from './Person/Person';
+
+const persons = (props) => props.persons.map((person, index) =>
+      (
+        <Person
+          click={() => { props.clicked(index); }}
+          name={person.name}
+          age={person.age}
+          key={person.id}
+          changed={(event) => { props.changed(event, person.id) }}
+        >
+          My Hobbies: Milruck
+          </Person>
+    ));
+
+export default persons;
